+++ src/main/java/org/binclassreader/structs/ConstClassInfo.java
@@ -26,7 +26,7 @@ import org.binclassreader.utils.Utilities;
 //https://docs.oracle.com/javase/specs/jvms/se8/html/jvms-4.html#jvms-4.4.1
 public class ConstClassInfo {
     @BinClassParser(byteToRead = 2)
-    private int[] name_index;
+    private short[] name_index;
 
     @PoolItemIndex(mustBeOfType = ConstUtf8Info.class)
     public int getNameIndex() {
